Bug fixes and code cleanup.

+++ Mhotivo/Controllers/CourseController.cs
@@ -16,14 +16,17 @@ namespace Mhotivo.Controllers
         private readonly ICourseRepository _courseRepository;
         private readonly IEducationLevelRepository _areaRepository;
         private readonly ViewMessageLogic _viewMessageLogic;
+        private IPensumRepository _pensumRepository;
 
         public CourseController(ICourseRepository courseRepository, 
-                                IEducationLevelRepository areaRepository)
+                                IEducationLevelRepository areaRepository,
+                                IPensumRepository pensumRepository)
         {
             if (courseRepository == null) throw new ArgumentNullException("courseRepository");
             if (areaRepository == null) throw new ArgumentNullException("areaRepository");
             _courseRepository = courseRepository;
             _areaRepository = areaRepository;
+            _pensumRepository = pensumRepository;
             _viewMessageLogic = new ViewMessageLogic(this);
         }
 
@@ -116,11 +119,22 @@ namespace Mhotivo.Controllers
         [AuthorizeAdmin]
         public ActionResult Delete(long id)
         {
-            var course = _courseRepository.Delete(id);
-            const string title = "Materia Eliminada";
-            var content = course.Name + " ha sido eliminado exitosamente.";
-            _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.InformationMessage);
-            return RedirectToAction("Index");
+            var check = _pensumRepository.Filter(x => x.Course.Id == id).FirstOrDefault();
+            if (check == null)
+            {
+                var course = _courseRepository.Delete(id);
+                const string title = "Materia Eliminada";
+                var content = course.Name + " ha sido eliminado exitosamente.";
+                _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.SuccessMessage);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                const string title = "Error!";
+                var content = "No se puede borrar la materia pues existe un pensum con esta materia.";
+                _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.ErrorMessage);
+                return RedirectToAction("Index");
+            }
         }
         
         /// GET: /Course/Edit
