+++ Mhotivo/Controllers/GradeController.cs
@@ -15,11 +15,13 @@ namespace Mhotivo.Controllers
     {
         private readonly IGradeRepository _gradeRepository;
         private readonly ViewMessageLogic _viewMessageLogic;
+        private IAcademicYearRepository _academicYearRepository;
 
-        public GradeController(IGradeRepository gradeRepository)
+        public GradeController(IGradeRepository gradeRepository, IAcademicYearRepository academicYearRepository)
         {
             if (gradeRepository == null) throw new ArgumentNullException("gradeRepository");
             _gradeRepository = gradeRepository;
+            _academicYearRepository = academicYearRepository;
             _viewMessageLogic = new ViewMessageLogic(this);
         }
 
@@ -99,11 +101,22 @@ namespace Mhotivo.Controllers
         [AuthorizeAdmin]
         public ActionResult Delete(long id)
         {
-            var grade = _gradeRepository.Delete(id);
-            const string title = "Grado ha sido Eliminado";
-            var content = grade.Name + " ha sido eliminado exitosamente.";
-            _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.InformationMessage);
-            return RedirectToAction("Index");
+            var check = _academicYearRepository.Filter(x => x.Grade.Id == id).FirstOrDefault();
+            if (check == null)
+            {
+                var grade = _gradeRepository.Delete(id);
+                const string title = "Grado ha sido Eliminado";
+                var content = grade.Name + " ha sido eliminado exitosamente.";
+                _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.SuccessMessage);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                const string title = "Error!";
+                var content = "No se puede borrar el grado pues existe un año académico con este grado.";
+                _viewMessageLogic.SetNewMessage(title, content, ViewMessageType.ErrorMessage);
+                return RedirectToAction("Index");
+            }
         }
         
         /// GET: /Grade/Edit/5
